sync - Sep18
current branch used in tilde-imgui

+++ pd/cpd/cpd_array_element.h
@@ -55,7 +55,7 @@ CPD_EXTERN t_cpd_array_element* cpd_element_prev(t_cpd_array_element* el);
  * @return new pointer
  * @note no NULL check
  */
-CPD_EXTERN t_cpd_array_element* cpd_element_move(t_cpd_array_element* el, ssize_t offset);
+CPD_EXTERN t_cpd_array_element* cpd_element_move(t_cpd_array_element* el, long offset);
 
 /**
  * Returns element difference
